split_solve represents services w multiple points
representative_vrp represents the original VRP by merging all points
belonging to the same vehicle_cluster (coming from the first split) in
to several points -- one is the mean and the remaining are the extreme
points of all services of vehicle v. This way clustering can take into
account the distribution of the original services.