Fix request_helper shadowing failing tests
last_response.body calls a method in rack/mock.rb:215 which casts
the message into a string; however, somethimes (when the moon and starts
align) response body is a proper Api::V01::VrpResult object
and mock (understandably) fails to convert it to a String.

Bottomline this error hides the main reason why the test fails -- which
is wrong return status.